mvc method started
controllers and views added
