+++ express/controller/errorController.js
@@ -1,0 +1,4 @@
+exports.get404Page = (req, res, next) =>{
+    res.status(404);
+    res.render('error', { pageTitle : 'page not found', path : ''});
+};
