+++ express/controller/products.js
@@ -1,0 +1,15 @@
+const products = [];
+
+exports.getAddProduct = (req, res, next) => {
+    res.status(200);
+    res.render('add-product', { pageTitle : 'Add product', path : '/admin/add-product'});
+};
+
+exports.postAddproduct = (req, res, next) => {
+    products.push({title: req.body.title});
+    res.redirect('/');
+};
+
+exports.getProducts = (req, res, next) => {
+    res.render('shop', {prods: products, pageTitle : 'shop', path : '/'});
+};
