+++ express/routes/admin.js
@@ -1,24 +1,17 @@
 const express = require('express');
-const path = require('path');
-const rootDir = require('../util/path');
+
+const productController = require('../controller/products');
 
 
 const router = express.Router();
-const products = [];
+
 
 // it is reached '/admin/add-product => GET
-router.get('/add-product', (req, res, next) => {
-    res.status(200);
-    //res.sendFile(path.join(rootDir, 'views', 'add-product.html')); //to get path of the html file
-    res.render('add-product', { pageTitle : 'Add product', path : '/admin/add-product'});
-});
+router.get('/add-product', productController.getAddProduct);
 
 // it is reached '/admin/add-product => POST
-router.post('/add-product', (req, res, next) => {
-    products.push({title: req.body.title});
-    res.redirect('/');
-});
+router.post('/add-product', productController.postAddproduct);
 
 exports.router = router;
-exports.products = products;
+
