+++ express/routes/error404.js
@@ -1,13 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-const path = require('path');
+const errorController = require('../controller/errorController');
 
-router.use((req, res, next) =>{
-    res.status(404);
-    //res.send("<h1>page not found!!</h1>");
-    //res.sendFile(path.join(__dirname, '../', 'views', 'error.html')); //to get path of the html file
-    res.render('error', { pageTitle : 'page not found'});
-});
+router.use(errorController.get404Page);
 
 module.exports = router;
