+++ express/routes/shop.js
@@ -1,12 +1,8 @@
 const express = require('express');
-const path = require('path');
-const rootDir = require('../util/path');
 
-const products = require('./admin');
+const productController = require('../controller/products');
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-    res.render('shop', {prods: products.products, pageTitle : 'shop', path : '/'});
-});
+router.get('/', productController.getProducts);
 
 module.exports = router;
